Move ArbitraryValidator to api module
