+++ fixture-monkey-api/build.gradle
@@ -9,6 +9,7 @@ plugins {
 
 dependencies {
     api("org.apiguardian:apiguardian-api:1.1.2")
+    api("javax.validation:validation-api:2.0.1.Final")
     compileOnly("net.jqwik:jqwik-engine:${JQWIK_VERSION}")
     compileOnly("net.jqwik:jqwik-api:${JQWIK_VERSION}")
     compileOnly("net.jqwik:jqwik-time:${JQWIK_VERSION}")
