--- fixture-monkey-mockito/src/main/java/com/navercorp/fixturemonkey/mockito/arbitrary/MockitoInterfaceSupplier.java
@@ -1,35 +1,0 @@
-/*
- * Fixture Monkey
- *
- * Copyright (c) 2021-present NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.navercorp.fixturemonkey.mockito.arbitrary;
-
-import org.mockito.Mockito;
-
-import com.navercorp.fixturemonkey.arbitrary.InterfaceSupplier;
-
-public final class MockitoInterfaceSupplier<T> implements InterfaceSupplier<T> {
-	public static final MockitoInterfaceSupplier<?> INSTANCE = new MockitoInterfaceSupplier<>();
-
-	private MockitoInterfaceSupplier() {
-	}
-
-	@Override
-	public T get(Class<T> type) {
-		return Mockito.mock(type);
-	}
-}
